requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Database & ORM
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Data Visualization & Dashboard
streamlit>=1.28.0
plotly>=5.17.0

# Data Sources
kagglehub>=0.2.0

# Environment & Configuration
python-dotenv>=1.0.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Optional: Enhanced Analytics
scikit-learn>=1.3.0
scipy>=1.11.0

